Refactor a test case file
- Rename test case class and file name
- Remove XCTUnwrap
- Remove unnecessary knowledge

cf. https://github.com/YusukeHosonuma/SwiftPrettyPrint/pull/155#discussion_r571964638
https://github.com/YusukeHosonuma/SwiftPrettyPrint/pull/155#discussion_r571964638
